pytest

+++ Project/Fast_API/planner/routes/events.py
@@ -1,10 +1,13 @@
 from typing import List
 
 from beanie import PydanticObjectId
+from bson import ObjectId
+
 from fastapi import APIRouter, Depends, HTTPException, Request, status
 from database.connection import get_session, Database
 from sqlmodel import select
 
+from auth.authenticate import authenticate
 from models.events import Event, EventUpdate
 
 event_router = APIRouter(
@@ -92,6 +95,68 @@ events = []
 
 #==================================  beanie(mongoDB)  ======================================
 
+# event_database = Database(Event)
+
+# # 전체 이벤트 조회
+# @event_router.get("/", response_model=List[Event])
+# async def retrieve_all_events() -> List[Event]:
+#     events = await event_database.get_all()
+#     return events
+
+# # 단일 이벤트 조회
+# @event_router.get("/{id}", response_model=Event)
+# async def retrieve_event(id: PydanticObjectId) -> Event:
+#     event = await event_database.get(id)
+#     if not event:
+#         raise HTTPException(
+#             status_code=status.HTTP_404_NOT_FOUND,
+#             detail="Event with supplied ID does not exist"
+#         )   
+#     return event
+
+# # 이벤트 저장
+# @event_router.post("/new")
+# async def create_event(body: Event) -> dict:
+#         await event_database.save(body)    
+#         return {
+#         "message": "Event created successfully"
+#     }
+
+# # 이벤트 수정
+# @event_router.put("/{id}", response_model=Event)
+# async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+#     updated_event = await event_database.update(id, body)
+#     if not updated_event:        
+#         raise HTTPException(
+#             status_code=status.HTTP_404_NOT_FOUND,
+#             detail="Event with supplied ID does not exist"
+#         )
+#     return updated_event
+
+# # 단일 이벤트 삭제
+# @event_router.delete("/{id}")
+# async def delete_event(id: PydanticObjectId) -> dict:
+#     event = await event_database.delete(id)
+#     if not event:        
+#         raise HTTPException(
+#             status_code=status.HTTP_404_NOT_FOUND,
+#             detail="Event with supplied ID does not exist"
+#         )
+#     return {
+#         "message": "Event deleted successfully."
+#     }
+  
+# # 전체 이벤트 삭제
+# @event_router.delete("/")
+# async def delete_all_events() -> dict:
+#     events.clear()
+#     return {
+#         "message": "Events deleted successfully"
+#     }
+
+
+#==================================  authenticate  ======================================
+
 event_database = Database(Event)
 
 # 전체 이벤트 조회
@@ -113,40 +178,54 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 # 이벤트 저장
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
-        await event_database.save(body)    
-        return {
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creater = user
+    await event_database.save(body)
+    return {
         "message": "Event created successfully"
     }
 
+
 # 이벤트 수정
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    # print(f"Retrieved event: {id}, Type: {type(id)}")
+
+    event = await event_database.get(id)
+    print(f"Retrieved event: {event}, Type: {type(event)}")
+
+
+    if event.creater != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
-    if not updated_event:        
+    if not updated_event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
     return updated_event
 
+
+
 # 단일 이벤트 삭제
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
-    if not event:        
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist"
+            detail="Event not found"
         )
+    if event.creater != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
+    event = await event_database.delete(id)
+
     return {
         "message": "Event deleted successfully."
-    }
-  
-# 전체 이벤트 삭제
-@event_router.delete("/")
-async def delete_all_events() -> dict:
-    events.clear()
-    return {
-        "message": "Events deleted successfully"
     }
